backend/requirements.txt
# FastAPI Web Framework
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-multipart==0.0.9
gunicorn==21.2.0

# Google Cloud & Firebase
google-cloud-vision==3.7.2
google-cloud-bigquery==3.21.0
google-cloud-storage==2.16.0
google-cloud-firestore==2.16.0
google-cloud-aiplatform==1.49.0
firebase-admin==6.5.0

# Document Parsers
PyMuPDF==1.24.1
python-docx==1.1.2
numpy==1.26.4